Don't crash after deleting a note
The problem is that `deleteNote` gets rid of only the note level
reference while lane reference to note remains. For now this has been
fixed by making `Lanes` logic more robust. It simply skips notes that
have been deleted.

Instead, it might be a better idea to get rid of the note from the lane
in question as well. This would require some extra logic and an action.

Closes #4.